Particle: State to class enum

+++ src/particle/src/engine/4C_particle_engine_container.hpp
@@ -169,7 +169,7 @@ namespace Particle
         FOUR_C_THROW("particle state '{}' not stored in container!", enum_to_state_name(state));
 #endif
 
-      return statedim_[state];
+      return statedim_[static_cast<int>(state)];
     };
 
     //! \name access global id and particle states
@@ -305,7 +305,8 @@ namespace Particle
 
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[state]); ++i) state_ptr[i] *= fac;
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(state)]); ++i)
+        state_ptr[i] *= fac;
     };
 
     /*!
@@ -329,7 +330,7 @@ namespace Particle
       if (not storedstates_.contains(stateB))
         FOUR_C_THROW("particle state '{}' not stored in container!", enum_to_state_name(stateB));
 
-      if (statedim_[stateA] != statedim_[stateB])
+      if (statedim_[static_cast<int>(stateA)] != statedim_[static_cast<int>(stateB)])
         FOUR_C_THROW("dimensions of states do not match!");
 #endif
 
@@ -338,7 +339,7 @@ namespace Particle
       const double* state_b_ptr = get_ptr_to_state(stateB, 0);
       double* state_a_ptr = get_ptr_to_state_writable(stateA, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[stateA]); ++i)
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(stateA)]); ++i)
         state_a_ptr[i] = facA * state_a_ptr[i] + facB * state_b_ptr[i];
     };
 
@@ -355,7 +356,7 @@ namespace Particle
       if (not storedstates_.contains(state))
         FOUR_C_THROW("particle state '{}' not stored in container!", enum_to_state_name(state));
 
-      if (statedim_[state] != static_cast<int>(val.size()))
+      if (statedim_[static_cast<int>(state)] != static_cast<int>(val.size()))
         FOUR_C_THROW("dimensions of states do not match!");
 #endif
 
@@ -364,8 +365,8 @@ namespace Particle
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
       for (int i = 0; i < particlestored_; ++i)
-        for (int dim = 0; dim < statedim_[state]; ++dim)
-          state_ptr[i * statedim_[state] + dim] = val[dim];
+        for (int dim = 0; dim < statedim_[static_cast<int>(state)]; ++dim)
+          state_ptr[i * statedim_[static_cast<int>(state)] + dim] = val[dim];
     };
 
     /*!
@@ -385,7 +386,8 @@ namespace Particle
 
       double* state_ptr = get_ptr_to_state_writable(state, 0);
 
-      for (int i = 0; i < (particlestored_ * statedim_[state]); ++i) state_ptr[i] = 0.0;
+      for (int i = 0; i < (particlestored_ * statedim_[static_cast<int>(state)]); ++i)
+        state_ptr[i] = 0.0;
     };
 
     //! @}
